requirements.txt
asgiref==3.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.42.0
django-bootstrap4==2.3.1
django-braces==1.14.0
django-crispy-forms==1.9.2
django-extra-views==0.13.0
django-heroku==0.3.1
django-registration==3.1
django-rest-auth==0.9.5
django-webpack-loader==0.6.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.10
mysqlclient==2.0.1
oauthlib==3.1.0
Pillow==8.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
